// ItemsAndSpells/ItemTypes/Types.tsx
export enum Types{
    WhiteMagic = "WHITE MAGIC",
    BlackMagic = "BLACK MAGIC",
    Staff = "STAFF",
    Sword = "SWORD",
    Bow = "BOW",
    Bracer = "BRACER",
    Hat = "HAT",
    HeavyArmor = "HEAVY ARMOR",
    Helm = "HELM",
    LightArmor = "LIGHT ARMOR",
    Robe = "ROBE",
    Shield = "SHIELD"
}